Add files via upload
this  contract works on the principle where the owner of the NFT will always be NFT’s owner. The person who will rent it will not be able to resell it, keep it or do anything else with it except for using it. It is the same situation when you rent an apartment in a building. You can live there as long as you have a lease contract but you don’t have the ownership and can’t sell it to other people or even destroy it  . I did this for the renting of the billboards ....